+++ core/src/main/kotlin/no/elg/hex/Settings.kt
@@ -70,7 +70,7 @@ object Settings {
   var enableActionHighlight by PreferenceDelegate(false, priority = 100)
   var enableDoubleTapToZoom by PreferenceDelegate(false, priority = 100)
   var showIslandId by PreferenceDelegate(false, priority = 110, shouldHide = { Hex.mapEditor || Hex.debug }) // always show island id in map editor/debug
-  var showArtbId by PreferenceDelegate(false, priority = 111, shouldHide = { Hex.mapEditor || Hex.debug }) // always show island id in map editor/debug
+  var showWorldRecord by PreferenceDelegate(false, priority = 111, shouldHide = { Hex.mapEditor || Hex.debug }) // always show island id in map editor/debug
 
   var enableStrengthHintEverywhere by PreferenceDelegate(false, priority = 220, shouldHide = { !Hex.debug && !Hex.mapEditor })
   var enableStrengthHintInPlayerTerritories by PreferenceDelegate(false, priority = 221, shouldHide = { !Hex.debug })